back/requirements.txt
django
django-rest-framework
psycopg2-binary
celery
django-cors-headers
